Parsed control set in AppCompatCache Parser
Context: The existing AppCompatCache Parser will parse all Controlsets
(seen in the test: TestFilters). This is validated in the `testFilters`
test which determines of the parser will parse Keys such as
HKLM\System\ControlSet001 and HKLM\System\Controlset47. In the commit
we have added a special field to report the control set, `control_set`.